Initial support for Azure Functions .NET 7 projects (isolated worker)
