+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/ReSharper.Azure/src/Azure.Intellisense/FunctionApp/ProjectTemplates/DotNetExtensions/Parameters/AzureFunctionsTemplateParameterProvider.cs
@@ -1,0 +1,43 @@
+// Copyright (c) 2022 JetBrains s.r.o.
+//
+// All rights reserved.
+//
+// MIT License
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+// documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
+// to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+// the Software.
+//
+// THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+// THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using System.Collections.Generic;
+using JetBrains.Application;
+using JetBrains.Rider.Backend.Features.ProjectModel.ProjectTemplates.DotNetExtensions;
+
+namespace JetBrains.ReSharper.Azure.Intellisense.FunctionApp.ProjectTemplates.DotNetExtensions.Parameters
+{
+    [ShellComponent]
+    public class AzureFunctionsTemplateParameterProvider : IDotNetTemplateParameterProvider
+    {
+        public int Priority => 20;
+
+        public IReadOnlyCollection<DotNetTemplateParameter> Get()
+        {
+            return new DotNetTemplateParameter[]
+            {
+                new AzureFunctionsVersionParameter(),
+                // TODO: When RdTextParameterStyle supports numeric/text only input, these can be enabled
+                //new AzureFunctionsHttpPortParameter(),
+                //new AzureFunctionsStorageConnectionStringParameter()
+            };
+        }
+    }
+}
