+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/resources/META-INF/plugin.xml
@@ -27,6 +27,7 @@
         <ul>
           <li>Compatibility with Rider 2022.2 EAP 8</li>
           <li>Azure Functions: Isolated functions project template is in net6-isolated folder in newer host versions (<a href="https://github.com/JetBrains/azure-tools-for-intellij/issues/599">#599</a>)</li>
+          <li>Azure Functions: Initial support for .NET 7 projects (isolated worker) (<a href="https://github.com/JetBrains/azure-tools-for-intellij/pull/615">#615</a>)</li>
         </ul>
         <h4>Fixed bugs:</h4>
         <ul>
